requirements.txt
# 动态记忆系统依赖

# 基础依赖
numpy>=1.20.0

# 讯飞星火大模型（必需）
websocket-client>=1.6.0

# LLM接口（可选，根据需要选择）
# OpenAI（可选）
# openai>=1.0.0

# Anthropic（可选）
# anthropic>=0.3.0

# 其他工具库（可选）
# 用于更好的JSON处理
# ujson>=5.0.0

# 用于进度条显示（可选）
# tqdm>=4.65.0
